backend-service/requirements.txt
ansible==4.5.0
ansible-core==2.11.5
apparmor==3.0.3
appdirs==1.4.4
arandr==0.1.10
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
binwalk==2.3.3
Brlapi==0.8.2
btrfsutil==5.14
CacheControl==0.12.6
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cloud-init==21.3
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==3.4.8
distlib==0.3.2
distro==1.6.0
Django==3.2.5
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
future==0.18.2
Glances==3.2.2
greenlet==1.1.1
gunicorn==20.1.0
html5lib==1.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
keyring==23.0.1
LibAppArmor==3.0.3
libfdt==1.6.1
libvirt-python==7.3.0
louis==3.19.0
MarkupSafe==2.0.1
more-itertools==8.9.0
msgpack==1.0.2
netifaces==0.11.0
nftables==0.1
numpy==1.21.1
oauthlib==3.1.1
ordered-set==4.0.2
packaging==20.9
paramiko==2.7.2
path==16.2.0
pep517==0.11.0
Pillow==8.3.1
ply==3.11
prettytable==2.2.0
progress==1.5
proton-client==0.5.1
protonvpn-cli==3.7.2
protonvpn-gui==1.0.1
protonvpn-nm-lib==3.3.2
psutil==5.8.0
pycairo==1.20.1
pycodestyle==2.7.0
pycparser==2.20
PyGObject==3.40.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.19.0
python-gnupg==0.4.7
pythondialog==3.5.2
pytube==11.0.1
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
ranger-fm==1.9.3
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.25
sqlparse==0.4.1
systemd-python==234
team==1.0
toml==0.10.2
tomli==1.2.1
ufw==0.36
urllib3==1.26.6
validate==5.1.0.dev0
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
youtube-dl==2021.6.6
zipp==3.5.0
